Remove int casting

+++ imfusion-plugin/lowdose-plugin/LowDoseSegmentationAlgorithm.h
@@ -18,7 +18,7 @@ namespace ImFusion
 		LowDoseSegmentationAlgorithm(SharedImageSet* img);
 
 		/// Set downsampling thickness
-		void setFactor(int thickness) { m_thickness = thickness; }
+		void setFactor(double thickness) { m_thickness = thickness; }
 
 		/// \name	Methods implementing the algorithm interface
 		//\{
@@ -41,6 +41,6 @@ namespace ImFusion
 	private:
 		SharedImageSet* m_imgIn = nullptr;           ///< Input image to process
 		std::unique_ptr<SharedImageSet> m_imgOut;    ///< Output image after processing
-		double m_thickness = 2;                            ///< Downsampling factor
+		double m_thickness = 2.0;                            ///< Downsampling factor
 	};
 }
